Rename CookieBanner to LocalStorageBanner and fix text
